backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
celery==5.4.0
redis==5.0.4
yt-dlp==2024.5.27
ffmpeg-python==0.2.0
python-multipart==0.0.9
requests==2.32.3
pydantic-settings==2.2.1
